refactor: 토큰 추출 로직 분리 (#19)
