+++ src/main/java/project/backend/presentation/auth/controller/AuthController.java
@@ -1,12 +1,8 @@
 package project.backend.presentation.auth.controller;
 
-import static project.backend.business.auth.implement.KakaoLoginManager.BEARER;
-
 import com.fasterxml.jackson.core.JsonProcessingException;
 import jakarta.servlet.http.HttpServletRequest;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -17,6 +13,7 @@ import org.springframework.web.bind.annotation.RestController;
 import project.backend.business.auth.AuthService;
 import project.backend.business.auth.request.TokenServiceRequest;
 import project.backend.business.auth.response.TokenServiceResponse;
+import project.backend.presentation.auth.util.TokenExtractor;
 
 @RestController
 @RequiredArgsConstructor
@@ -24,11 +21,7 @@ import project.backend.business.auth.response.TokenServiceResponse;
 public class AuthController {
 
   private final AuthService authService;
-
-  @Value("${jwt.access_header}")
-  private String accessTokenHeader;
-  @Value("${jwt.refresh_header}")
-  private String refreshTokenHeader;
+  private final TokenExtractor tokenExtractor;
 
   @RequestMapping("/login/kakao")
   public ResponseEntity<TokenServiceResponse> loginKakao(@RequestParam(name = "code") String code) throws JsonProcessingException {
@@ -38,33 +31,14 @@ public class AuthController {
 
   @GetMapping("/reissue")
   public ResponseEntity<TokenServiceResponse> reissueToken(HttpServletRequest request) {
-    TokenServiceRequest tokenServiceRequest = extractTokenRequest(request);
+    TokenServiceRequest tokenServiceRequest = tokenExtractor.extractTokenRequest(request);
     return new ResponseEntity<>(authService.reissueAccessToken(tokenServiceRequest), HttpStatus.OK);
   }
 
   @PostMapping("/logout")
   public ResponseEntity<Void> logout(HttpServletRequest request) {
-    TokenServiceRequest tokenServiceRequest = extractTokenRequest(request);
+    TokenServiceRequest tokenServiceRequest = tokenExtractor.extractTokenRequest(request);
     authService.logout(tokenServiceRequest);
     return ResponseEntity.ok().build();
-  }
-
-  private TokenServiceRequest extractTokenRequest(HttpServletRequest request) {
-    return TokenServiceRequest.builder()
-                              .accessToken(extractAccessToken(request).orElse(null))
-                              .refreshToken(extractRefreshToken(request).orElse(null))
-                              .build();
-  }
-
-  private Optional<String> extractAccessToken(HttpServletRequest request) {
-    return Optional.ofNullable(request.getHeader(accessTokenHeader))
-                   .filter(accessToken -> accessToken.startsWith(BEARER))
-                   .map(accessToken -> accessToken.replace(BEARER, ""));
-  }
-
-  private Optional<String> extractRefreshToken(HttpServletRequest request) {
-    return Optional.ofNullable(request.getHeader(refreshTokenHeader))
-                   .filter(refreshToken -> refreshToken.startsWith(BEARER))
-                   .map(refreshToken -> refreshToken.replace(BEARER, ""));
   }
 }
